Factor out execution report building logic.
