+++ ordermatch/src/test/java/quickfix/examples/ordermatch/MockMessageSender.java
@@ -1,4 +1,4 @@
-package quickfix.examples.utility;
+package quickfix.examples.ordermatch;
 
 import java.util.ArrayList;
 import java.util.List;
