+++ ordermatch/src/test/java/quickfix/examples/ordermatch/OrderMatchTest.java
@@ -11,7 +11,6 @@ import org.mockito.Mock;
 
 import quickfix.Message;
 import quickfix.SessionID;
-import quickfix.examples.utility.MockMessageSender;
 import quickfix.fix42.NewOrderSingle;
 
 public class OrderMatchTest {
